Fixed a bug when only a tail of an item is in view
When the view is moving to the right, an item might get partially
off-screen, but a character or two might still be on-screen. After last
commit, such items used to lose their styling.

Fixed by adding a special case to styling calculations, to account for
negative position of item start character.