Fixed the length of search highlight region
When searching for non-ASCII string, highlighting went wild due to
number of symbols being specified as highlight region, not the number of
bytes. Fixed that.

Also, testing this case revealed a bug in STC: when typing Unicode text,
the control refuses to take any more characters after you backspace.
A quick search revealed that this is a known bug in STC:
http://trac.wxwidgets.org/ticket/8300. It is supposed to be fixed, but
either my wxWidgets is too old, or the fix didn't help. Need more
investigation.