Added capability to refresh the view
This patch introduces the capability for the panel to refresh the
contents of the directory. From user's perspective, it's achieved by
pressing 'r'. The last two hunks in candy.py were exposed by the tests
and their nature is not entirely clear: somehow I manage to get None
objects for visualItems, and thus I put guards on them.