An ad-hoc functional test for cursor motion
This is a simple functional test rig, implemented on top of unittest,
but violating unit testing assumptions. In long term this should
suck up the "slow" test cases from the regular test suite.

Note: 6 out of 7 tests are failing, they expose current bugs in cursor
movement. This test suite was deliberately developed to expose and cover
that. Fixes for the code to follow.